application/requirements.txt
amqp==5.3.1
billiard==4.2.1
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
env==0.1.0
filelock==3.16.1
flask==3.1.0
fonttools==4.56.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
itsdangerous==2.2.0
jinja2==3.1.5
kiwisolver==1.4.7
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib==3.9.4
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
prompt-toolkit==3.0.50
pyparsing==3.2.1
pypdf2==3.0.1
python-dateutil==2.9.0.post0
reportlab==4.3.0
requests==2.32.3
six==1.17.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.27.0
wcwidth==0.2.13
werkzeug==3.1.3
windows-curses==2.4.1
zipp==3.21.0
